requirements.txt
facebook-sdk
ipdb